feat: some fix on subscriber

+++ src/helpers/subscribers.ts
@@ -1,4 +1,4 @@
-import { PrismaClient } from '@prisma/client'
+import { Interest, PrismaClient } from '@prisma/client'
 
 const prisma = new PrismaClient()
 
@@ -6,29 +6,40 @@ export const checkIfSameUser = function (id, idLoggedUser, res) {
   if (id !== idLoggedUser) res.status(403).json({ "error": "Você não está autorizado a fazer essa operação" })
 }
 
-export const bothConfirmation = async function (result) {
+export const bothConfirmation = async function (result: Interest) {
   if (result.pConfirmation && result.uConfirmation) {
-    const createRent = prisma.rent.create({
-      data: {
-        guest: {
-          connect: {
-            id: result.userId
-          }
-        },
-        property: {
-          connect: {
-            id: result.propertyId
-          }
-        }
+    const resultProperty = await prisma.property.findFirst({
+      where: {
+        id: result.propertyId
       }
     })
 
-    const deleteInterest = prisma.interest.deleteMany({
-      where: {
-        id: result.userId
-      }
-    })
-    const transactional = await prisma.$transaction([createRent, deleteInterest])
-    return transactional
+    //TEM QUE FAZER UM COUNT NO NUMERO DE RENTS E CHECAR
+
+    if (resultProperty.vacancyNumber > 0) {
+      const createRent = prisma.rent.create({
+        data: {
+          guest: {
+            connect: {
+              id: result.userId
+            }
+          },
+          property: {
+            connect: {
+              id: result.propertyId
+            }
+          }
+        }
+      })
+
+      const deleteInterest = prisma.interest.deleteMany({
+        where: {
+          id: result.userId
+        }
+      })
+      const transactional = await prisma.$transaction([createRent, deleteInterest])
+      return transactional
+    }
+
   }
 }
